agent-examples/05-langgraph-websearch-agent/requirements.txt
python-dotenv
langchain-community
langchain-core
langchain-openai
langgraph
google-search-results
